fix problems.

+++ src/dev_lib/smart_handle.ixx
@@ -1,7 +1,9 @@
-export module devlib.smart_handles;
+export module devlib.smart_handle;
 
 import std;
 import <cassert>;
+
+#define Do_Optimization 1
 
 namespace dev_lib {
     export struct atomic_ref_count_info_type;
@@ -92,7 +94,11 @@ namespace dev_lib {
         using handle_type = t_handle_type;
 
     private:
-        std::atomic<control_block_type *> m_control_block{nullptr};
+#if Do_Optimization
+        mutable std::atomic<control_block_type *> m_control_block{nullptr};
+#else
+        control_block_type *m_control_block;
+#endif
         std::optional<handle_type> m_handle{std::nullopt};
 
         strong_arc_handle(control_block_type *cb, handle_type handle) noexcept
@@ -101,12 +107,33 @@ namespace dev_lib {
         friend class weak_arc_handle<t_handle_type, t_info_type>;
 
     public:
+#if Do_Optimization
         strong_arc_handle() noexcept = default;
+#else
+        strong_arc_handle() noexcept
+            : m_control_block() {
+            allocator_type allocator;
+            control_block_type *new_cb = std::allocator_traits<allocator_type>::allocate(allocator, 1);
+            std::allocator_traits<allocator_type>::construct(allocator, new_cb);
+            m_control_block = new_cb;
+        }
+#endif
 
+#if Do_Optimization
         strong_arc_handle(handle_type handle) noexcept
             : m_handle(handle) {}
+#else
+        strong_arc_handle(handle_type handle) noexcept
+            : m_handle(handle) {
+            allocator_type allocator;
+            control_block_type *new_cb = std::allocator_traits<allocator_type>::allocate(allocator, 1);
+            std::allocator_traits<allocator_type>::construct(allocator, new_cb);
+            m_control_block = new_cb;
+        }
+#endif
 
         ~strong_arc_handle() noexcept {
+#if Do_Optimization
             auto cb = m_control_block.load(std::memory_order_relaxed);
             if (!cb) {
                 // No control block, means the control block is never created, check if handle exists
@@ -116,9 +143,12 @@ namespace dev_lib {
 
                 return;
             }
+#else
+            auto cb = m_control_block;
+#endif
 
             // Release strong reference
-            if (cb->release_strong_ref()) {
+            if (cb->release_strong_ref() && m_handle.has_value()) {
                 // Last strong reference released, destroy handle
                 info_type::destroy_handle(std::exchange(m_handle, std::nullopt).value());
             }
@@ -130,6 +160,8 @@ namespace dev_lib {
                 std::allocator_traits<allocator_type>::deallocate(allocator, cb, 1);
             }
         }
+
+#if Do_Optimization
 
         strong_arc_handle(strong_arc_handle &other) {
             auto other_cb = other.m_control_block.load(std::memory_order_relaxed);
@@ -161,17 +193,36 @@ namespace dev_lib {
             // else both are empty, do nothing
         }
 
+#else
+        strong_arc_handle(const strong_arc_handle &other)
+            : m_control_block(other.m_control_block),
+              m_handle(other.m_handle) {
+            if (m_control_block) {
+                m_control_block->add_strong_ref();
+                m_control_block->add_weak_ref();
+            }
+        }
+#endif
+
         strong_arc_handle(strong_arc_handle &&other) noexcept
             : m_control_block(
                   other.m_control_block.exchange(
                       nullptr, std::memory_order_relaxed)),
               m_handle(std::exchange(other.m_handle, std::nullopt)) {}
 
+#if Do_Optimization
         strong_arc_handle &operator=(strong_arc_handle &other) noexcept {
             this->~strong_arc_handle();
             new(this) strong_arc_handle(other);
             return *this;
         }
+#else
+        strong_arc_handle &operator=(const strong_arc_handle &other) noexcept {
+            this->~strong_arc_handle();
+            new(this) strong_arc_handle(other);
+            return *this;
+        }
+#endif
 
         strong_arc_handle &operator=(strong_arc_handle &&other) noexcept {
             this->~strong_arc_handle();
@@ -179,8 +230,16 @@ namespace dev_lib {
             return *this;
         }
 
-        handle_type &get() noexcept {
+        handle_type get() const noexcept {
             return m_handle.value();
+        }
+
+        const handle_type &operator*() const noexcept {
+            return m_handle.value(); // use safe access
+        }
+
+        const handle_type *operator->() const noexcept {
+            return &m_handle.value(); // use safe access
         }
 
         handle_type &operator*() noexcept {
@@ -195,10 +254,25 @@ namespace dev_lib {
             return m_handle.has_value();
         }
 
-        weak_arc_handle<t_handle_type, t_info_type> share_weak() noexcept;
-
+#if Do_Optimization
+        strong_arc_handle clone() const noexcept {
+            return strong_arc_handle(*this);
+        }
+#else
         strong_arc_handle clone() noexcept {
             return strong_arc_handle(*this);
+        }
+#endif
+
+#if Do_Optimization
+        weak_arc_handle<t_handle_type, t_info_type> share_weak() noexcept;
+#else
+        weak_arc_handle<t_handle_type, t_info_type> share_weak() const noexcept;
+#endif
+
+        void reset() noexcept {
+            this->~strong_arc_handle();
+            new(this) strong_arc_handle();
         }
     };
 
@@ -212,7 +286,7 @@ namespace dev_lib {
         using handle_type = t_handle_type;
 
     private:
-        std::atomic<control_block_type *> m_control_block{nullptr};
+        control_block_type *m_control_block{nullptr};
         std::optional<handle_type> m_handle{std::nullopt};
 
         weak_arc_handle(control_block_type *cb, handle_type handle) noexcept
@@ -223,11 +297,12 @@ namespace dev_lib {
     public:
         weak_arc_handle() noexcept = default;
 
+#if Do_Optimization
         weak_arc_handle(strong_arc_handle<t_handle_type, t_info_type> &strong_handle) noexcept {
             auto cb = strong_handle.m_control_block.load(std::memory_order_relaxed);
             if (cb) {
                 cb->add_weak_ref();
-                m_control_block.store(cb, std::memory_order_relaxed);
+                m_control_block = cb;
                 m_handle = strong_handle.m_handle;
                 return;
             }
@@ -240,7 +315,7 @@ namespace dev_lib {
 
                 new_cb->add_weak_ref();
 
-                m_control_block.store(new_cb, std::memory_order_relaxed);
+                m_control_block = new_cb;
                 auto old = strong_handle.m_control_block.exchange(new_cb, std::memory_order_relaxed);
                 assert(
                     old == nullptr &&
@@ -250,9 +325,22 @@ namespace dev_lib {
 
             // else both are empty, do nothing
         }
+#else
+        weak_arc_handle(const strong_arc_handle<t_handle_type, t_info_type> &strong_handle) noexcept {
+            auto cb = strong_handle.m_control_block;
+            if (!cb) {
+                assert(
+                    strong_handle.m_handle.has_value() == false);
+            }
+
+            cb->add_weak_ref();
+            m_control_block = cb;
+            m_handle = strong_handle.m_handle;
+        }
+#endif
 
         ~weak_arc_handle() noexcept {
-            auto cb = m_control_block.load(std::memory_order_relaxed);
+            auto cb = m_control_block;
             if (!cb) {
                 return;
             }
@@ -265,22 +353,23 @@ namespace dev_lib {
             }
         }
 
-        weak_arc_handle(weak_arc_handle &other) {
-            auto other_cb = other.m_control_block.load(std::memory_order_relaxed);
+        weak_arc_handle(const weak_arc_handle &other) {
+            auto other_cb = other.m_control_block;
             if (other_cb) {
                 other_cb->add_weak_ref();
-                m_control_block.store(other_cb, std::memory_order_relaxed);
+                m_control_block = other_cb;
                 m_handle = other.m_handle;
             }
         }
 
         weak_arc_handle(weak_arc_handle &&other) noexcept
             : m_control_block(
-                  other.m_control_block.exchange(
-                      nullptr, std::memory_order_relaxed)),
+                  std::exchange(
+                      other.m_control_block, nullptr)
+              ),
               m_handle(std::exchange(other.m_handle, std::nullopt)) {}
 
-        weak_arc_handle &operator=(weak_arc_handle &other) noexcept {
+        weak_arc_handle &operator=(const weak_arc_handle &other) noexcept {
             this->~weak_arc_handle();
             new(this) weak_arc_handle(other);
             return *this;
@@ -300,21 +389,36 @@ namespace dev_lib {
             return cb->get_strong_count().load(std::memory_order_acquire) == 0;
         }
 
-        strong_arc_handle<t_handle_type, t_info_type> lock();
+        strong_arc_handle<t_handle_type, t_info_type> lock() const;
 
-        weak_arc_handle clone() noexcept {
+        weak_arc_handle clone() const noexcept {
             return weak_arc_handle(*this);
+        }
+
+        void reset() noexcept {
+            this->~weak_arc_handle();
+            new(this) weak_arc_handle();
         }
     };
 
+#if Do_Optimization
     template<typename t_handle_type, typename t_info_type> requires std::is_trivially_copyable_v<t_handle_type>
-    weak_arc_handle<t_handle_type, t_info_type> strong_arc_handle<t_handle_type, t_info_type>::share_weak() noexcept {
+    weak_arc_handle<t_handle_type, t_info_type> strong_arc_handle<t_handle_type,
+        t_info_type>::share_weak() noexcept {
         return weak_arc_handle<t_handle_type, t_info_type>(*this);
     }
+#else
+    template<typename t_handle_type, typename t_info_type> requires std::is_trivially_copyable_v<t_handle_type>
+    weak_arc_handle<t_handle_type, t_info_type> strong_arc_handle<t_handle_type,
+        t_info_type>::share_weak() const noexcept {
+        return weak_arc_handle<t_handle_type, t_info_type>(*this);
+    }
+#endif
+
 
     template<typename t_handle_type, typename t_info_type> requires std::is_trivially_copyable_v<t_handle_type>
-    strong_arc_handle<t_handle_type, t_info_type> weak_arc_handle<t_handle_type, t_info_type>::lock() {
-        auto cb = m_control_block.load(std::memory_order_relaxed);
+    strong_arc_handle<t_handle_type, t_info_type> weak_arc_handle<t_handle_type, t_info_type>::lock() const {
+        auto cb = m_control_block;
         if (!cb) {
             return strong_arc_handle<t_handle_type, t_info_type>();
         }
@@ -400,7 +504,7 @@ namespace dev_lib {
 
     export template<typename t_handle_type, typename t_info_type = ref_count_info_type>
         requires std::is_trivially_copyable_v<t_handle_type>
-    class shared_rc_handle;
+    class strong_rc_handle;
 
     export template<typename t_handle_type, typename t_info_type = ref_count_info_type>
         requires std::is_trivially_copyable_v<t_handle_type>
@@ -408,7 +512,7 @@ namespace dev_lib {
 
     export template<typename t_handle_type, typename t_info_type>
         requires std::is_trivially_copyable_v<t_handle_type>
-    class shared_rc_handle {
+    class strong_rc_handle {
     public:
         using info_type = t_info_type;
         using control_block_type = reference_counter_control_block<info_type>;
@@ -416,21 +520,21 @@ namespace dev_lib {
         using handle_type = t_handle_type;
 
     private:
-        control_block_type *m_control_block{nullptr};
+        control_block_type mutable *m_control_block{nullptr};
         std::optional<handle_type> m_handle{std::nullopt};
 
-        shared_rc_handle(control_block_type *cb, handle_type handle) noexcept
+        strong_rc_handle(control_block_type *cb, handle_type handle) noexcept
             : m_control_block(cb), m_handle(handle) {}
 
         friend class weak_rc_handle<t_handle_type, t_info_type>;
 
     public:
-        shared_rc_handle() noexcept = default;
+        strong_rc_handle() noexcept = default;
 
-        shared_rc_handle(handle_type handle) noexcept
+        strong_rc_handle(handle_type handle) noexcept
             : m_handle(handle) {}
 
-        ~shared_rc_handle() noexcept {
+        ~strong_rc_handle() noexcept {
             if (!m_control_block) {
                 // No control block, means the control block is never created, check if handle exists
                 if (m_handle.has_value()) {
@@ -440,7 +544,7 @@ namespace dev_lib {
             }
 
             // Release strong reference
-            if (m_control_block->release_strong_ref()) {
+            if (m_control_block->release_strong_ref() && m_handle.has_value()) {
                 // Last strong reference released, destroy handle
                 info_type::destroy_handle(std::exchange(m_handle, std::nullopt).value());
             }
@@ -453,7 +557,7 @@ namespace dev_lib {
             }
         }
 
-        shared_rc_handle(shared_rc_handle &other) {
+        strong_rc_handle(const strong_rc_handle &other) {
             if (other.m_control_block) {
                 other.m_control_block->add_strong_ref();
                 other.m_control_block->add_weak_ref();
@@ -482,21 +586,21 @@ namespace dev_lib {
             // else both are empty, do nothing
         }
 
-        shared_rc_handle(shared_rc_handle &&other) noexcept
+        strong_rc_handle(strong_rc_handle &&other) noexcept
             : m_control_block(
                   std::exchange(
                       other.m_control_block, nullptr)),
               m_handle(std::exchange(other.m_handle, std::nullopt)) {}
 
-        shared_rc_handle &operator=(shared_rc_handle &other) noexcept {
+        strong_rc_handle &operator=(const strong_rc_handle &other) noexcept {
             this->~shared_rc_handle();
-            new(this) shared_rc_handle(other);
+            new(this) strong_rc_handle(other);
             return *this;
         }
 
-        shared_rc_handle &operator=(shared_rc_handle &&other) noexcept {
+        strong_rc_handle &operator=(strong_rc_handle &&other) noexcept {
             this->~shared_rc_handle();
-            new(this) shared_rc_handle(std::move(other));
+            new(this) strong_rc_handle(std::move(other));
             return *this;
         }
 
@@ -516,9 +620,9 @@ namespace dev_lib {
             return m_handle.has_value();
         }
 
-        weak_rc_handle<t_handle_type, t_info_type> share_weak() noexcept;
+        weak_rc_handle<t_handle_type, t_info_type> share_weak() const noexcept;
 
-        shared_rc_handle clone() noexcept {
+        strong_rc_handle clone() const noexcept {
             return shared_rc_handle(*this);
         }
     };
@@ -539,12 +643,12 @@ namespace dev_lib {
         weak_rc_handle(control_block_type *cb, handle_type handle) noexcept
             : m_control_block(cb), m_handle(handle) {}
 
-        friend class shared_rc_handle<t_handle_type, t_info_type>;
+        friend class strong_rc_handle<t_handle_type, t_info_type>;
 
     public:
         weak_rc_handle() noexcept = default;
 
-        weak_rc_handle(shared_rc_handle<t_handle_type, t_info_type> &shared_handle) noexcept {
+        weak_rc_handle(const strong_rc_handle<t_handle_type, t_info_type> &shared_handle) noexcept {
             if (shared_handle.m_control_block) {
                 shared_handle.m_control_block->add_weak_ref();
                 m_control_block = shared_handle.m_control_block;
@@ -584,7 +688,7 @@ namespace dev_lib {
             }
         }
 
-        weak_rc_handle(weak_rc_handle &other) {
+        weak_rc_handle(const weak_rc_handle &other) {
             if (other.m_control_block) {
                 other.m_control_block->add_weak_ref();
                 m_control_block = other.m_control_block;
@@ -598,7 +702,7 @@ namespace dev_lib {
                       other.m_control_block, nullptr)),
               m_handle(std::exchange(other.m_handle, std::nullopt)) {}
 
-        weak_rc_handle &operator=(weak_rc_handle &other) noexcept {
+        weak_rc_handle &operator=(const weak_rc_handle &other) noexcept {
             this->~weak_rc_handle();
             new(this) weak_rc_handle(other);
             return *this;
@@ -617,25 +721,37 @@ namespace dev_lib {
             return m_control_block->get_strong_count() == 0;
         }
 
-        shared_rc_handle<t_handle_type, t_info_type> lock();
+        strong_rc_handle<t_handle_type, t_info_type> lock() const;
 
-        weak_rc_handle clone() noexcept {
+        weak_rc_handle clone() const noexcept {
             return weak_rc_handle(*this);
         }
     };
 
     template<typename t_handle_type, typename t_info_type> requires std::is_trivially_copyable_v<t_handle_type>
-    shared_rc_handle<t_handle_type, t_info_type> weak_rc_handle<t_handle_type, t_info_type>::lock() {
+    weak_rc_handle<t_handle_type, t_info_type> strong_rc_handle<t_handle_type, t_info_type>::
+    share_weak() const noexcept {
+        return weak_rc_handle<t_handle_type, t_info_type>(*this);
+    }
+
+    template<typename t_handle_type, typename t_info_type> requires std::is_trivially_copyable_v<t_handle_type>
+    strong_rc_handle<t_handle_type, t_info_type> weak_rc_handle<t_handle_type, t_info_type>::lock() const {
         if (!m_control_block) {
-            return shared_rc_handle<t_handle_type, t_info_type>();
+            return strong_rc_handle<t_handle_type, t_info_type>();
         }
 
         auto locked_cb = m_control_block->lock_from_weak();
         if (!locked_cb) {
-            return shared_rc_handle<t_handle_type, t_info_type>();
+            return strong_rc_handle<t_handle_type, t_info_type>();
         }
 
-        return shared_rc_handle<t_handle_type, t_info_type>(locked_cb, m_handle.value());
+        return strong_rc_handle<t_handle_type, t_info_type>(locked_cb, m_handle.value());
+    }
+
+    export template<typename t_handle_type, typename t_info_type = ref_count_info_type, typename... Args>
+        requires std::is_trivially_copyable_v<t_handle_type>
+    auto make_shared_rc(Args &&... args) {
+        return strong_rc_handle<t_handle_type, t_info_type>(t_handle_type(std::forward<Args>(args)...));
     }
 
     // unique handle
@@ -680,8 +796,16 @@ namespace dev_lib {
             return *this;
         }
 
-        handle_type &get() noexcept {
+        handle_type get() const noexcept {
             return m_handle.value();
+        }
+
+        const handle_type &operator*() const noexcept {
+            return m_handle.value(); // use safe access
+        }
+
+        const handle_type *operator->() const noexcept {
+            return &m_handle.value(); // use safe access
         }
 
         handle_type &operator*() noexcept {
@@ -705,4 +829,10 @@ namespace dev_lib {
             return std::exchange(m_handle, std::nullopt).value();
         }
     };
+
+    export template<typename t_handle_type, typename t_info_type = unique_handle_info, typename... Args>
+        requires std::is_trivially_copyable_v<t_handle_type>
+    auto make_unique_handle(Args &&... args) {
+        return unique_handle<t_handle_type, t_info_type>(t_handle_type(std::forward<Args>(args)...));
+    }
 }
